refactor: Product를 Dao에서 JPA를 사용하도록 변경 및 패키지 이동
