+++ src/main/java/kitchenpos/product/appliction/ProductService.java
@@ -1,7 +1,8 @@
-package kitchenpos.application;
+package kitchenpos.product.appliction;
 
-import kitchenpos.dao.ProductDao;
+import kitchenpos.product.dao.ProductDao;
 import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,17 +12,17 @@ import java.util.Objects;
 
 @Service
 public class ProductService {
-    private final ProductDao productDao;
+    private final ProductRepository productRepository;
 
-    public ProductService(final ProductDao productDao) {
-        this.productDao = productDao;
+    public ProductService(ProductRepository productRepository) {
+        this.productRepository = productRepository;
     }
 
     @Transactional
     public Product create(final Product product) {
         this.validateProductByPrice(product);
 
-        return productDao.save(product);
+        return this.productRepository.save(product);
     }
 
     /**
@@ -37,6 +38,6 @@ public class ProductService {
     }
 
     public List<Product> list() {
-        return productDao.findAll();
+        return this.productRepository.findAll();
     }
 }
