+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,8 +1,14 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
+import javax.persistence.*;
 import java.math.BigDecimal;
 
+@Entity
+@Table(name = "product")
 public class Product {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
     private BigDecimal price;
