+++ src/test/java/kitchenpos/menu/MenuGroupControllerTest.java
@@ -3,14 +3,11 @@ package kitchenpos.menu;
 import com.fasterxml.jackson.core.type.TypeReference;
 import kitchenpos.common.BaseContollerTest;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Product;
-import kitchenpos.product.ProductTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MvcResult;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 
 import static org.assertj.core.api.Assertions.assertThat;
